requirements.txt
Pillow
numpy
face-recognition
opencv-python
transformers
tensorflow
pytorch